requirements-from-project.txt
aioboto3==10.1.0
aiobotocore==2.4.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.6.0
asn1crypto==1.4.0
astroid==2.5
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
awswrangler==2.9.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
black==21.5b0
boto3==1.24.59
botocore==1.27.59
cachecontrol==0.12.11
cachy==0.3.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==3.0.1
cleo==0.8.1
click==8.1.3
clikit==0.6.2
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.3.1
cryptography==39.0.1
distlib==0.3.6
django-computed-property==0.3.0
django-cors-headers==3.2.1
django-filter==2.4.0
django==3.0.3
djangorestframework-simplejwt==4.4.0
djangorestframework==3.11.0
drf-yasg==1.17.1
et-xmlfile==1.1.0
filelock==3.9.0
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.0.4
html5lib==1.1
identify==2.5.18
idna==2.10
inflection==0.5.1
isort==5.8.0
itypes==1.2.0
jeepney==0.8.0
jinja2==3.1.2
jmespath==1.0.1
keyring==21.8.0
lazy-object-proxy==1.9.0
lockfile==0.12.2
lxml==4.9.2
markupsafe==2.1.2
mccabe==0.6.1
msgpack==1.0.4
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.20.2
openpyxl==3.0.10
packaging==23.0
pandas==1.2.5
pandasql==0.7.3
paramiko==3.0.0
pastel==0.2.1
pathspec==0.11.0
pexpect==4.8.0
pg8000==1.19.5
pip==22.3.1
pkginfo==1.9.2
platformdirs==3.0.0
pre-commit==2.12.1
psycopg2==2.8.6
ptyprocess==0.7.0
pyarrow==4.0.0
pybase64==1.2.3
pycparser==2.21
pygments==2.9.0
pyjwt==1.7.1
pylev==1.4.0
pylint-django==2.3.0
pylint-plugin-utils==0.7
pylint==2.7.0
pymysql==1.0.2
pynacl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2020.4
pyyaml==6.0
redshift-connector==2.0.910
regex==2022.10.31
requests-toolbelt==0.9.1
requests==2.25.1
ruamel.yaml.clib==0.2.7
ruamel.yaml==0.17.21
s3transfer==0.6.0
scramp==1.4.0
secretstorage==3.3.3
setuptools==65.6.3
shellingham==1.5.0
six==1.16.0
soupsieve==2.4
sqlalchemy-utils==0.37.0
sqlalchemy==1.4.11
sqlparse==0.4.3
sshtunnel==0.4.0
toml==0.10.2
tomlkit==0.11.6
typing-extensions==4.5.0
uritemplate==4.1.1
uwsgi==2.0.18
virtualenv==20.19.0
webencodings==0.5.1
wheel==0.38.4
whitenoise==5.0.1
wrapt==1.12.1
xlrd==1.2.0
xmltodict==0.12.0
yarl==1.8.2